Rework: Split implementation into separate files
Move C files into src.
Move header files into include.
