+++ include/wolfengine.h
@@ -1,0 +1,102 @@
+/* wolfengine.h
+ *
+ * Copyright (C) 2019-2021 wolfSSL Inc.
+ *
+ * This file is part of wolfengine.
+ *
+ * wolfengine is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * wolfengine is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1335, USA
+ */
+
+#ifndef WOLFENGINE_H
+#define WOLFENGINE_H
+
+/* OpenSSL 3.0.0 has deprecated the ENGINE API. */
+#define OPENSSL_API_COMPAT      10101
+
+#include <openssl/engine.h>
+#include <openssl/evp.h>
+#include <openssl/ec.h>
+
+#include <wolfssl/options.h>
+#include <wolfssl/wolfcrypt/hash.h>
+#include <wolfssl/wolfcrypt/sha256.h>
+#include <wolfssl/wolfcrypt/aes.h>
+#include <wolfssl/wolfcrypt/ecc.h>
+
+#include "openssl_bc.h"
+
+#ifdef WOLFENGINE_DEBUG
+#define WOLFENGINE_MSG(msg)     (void)fprintf(stderr, "WOLFENG: %s\n", msg);
+#else
+#define WOLFENGINE_MSG(msg)     (void)msg
+#endif
+
+/*
+ * Global random
+ */
+
+extern WC_RNG* we_rng;
+
+/* For digest method in OpenSSL 1.0.2 */
+int we_pkey_get_nids(const int** nids);
+
+/*
+ * Digest methods.
+ */
+
+extern EVP_MD *we_sha256_md;
+int we_init_sha256_meth(void);
+
+extern EVP_MD *we_sha384_md;
+int we_init_sha384_meth(void);
+
+extern EVP_MD *we_sha512_md;
+int we_init_sha512_meth(void);
+
+extern EVP_MD *we_sha3_224_md;
+int we_init_sha3_224_meth(void);
+
+extern EVP_MD *we_sha3_256_md;
+int we_init_sha3_256_meth(void);
+
+extern EVP_MD *we_sha3_384_md;
+int we_init_sha3_384_meth(void);
+
+extern EVP_MD *we_sha3_512_md;
+int we_init_sha3_512_meth(void);
+
+/*
+ * Cipher methods.
+ */
+
+extern EVP_CIPHER* we_aes128_gcm_ciph;
+extern EVP_CIPHER* we_aes256_gcm_ciph;
+int we_init_aesgcm_meths(void);
+
+/*
+ * ECC methods.
+ */
+
+#ifdef WE_HAVE_EC_KEY
+extern EC_KEY_METHOD *we_ec_key_method;
+#endif
+extern EVP_PKEY_METHOD *we_ec_method;
+extern EVP_PKEY_METHOD *we_ec_p256_method;
+extern EVP_PKEY_METHOD *we_ec_p384_method;
+int we_init_ecc_meths(void);
+int we_init_ec_key_meths(void);
+
+#endif /* WOLFENGINE_H */
+
